# backend/api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/membership/update-subscription": {
      "post": {
        "tags": [
          "member-sub-controller"
        ],
        "operationId": "updateMemberSubscription",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonUpdateMemberSubscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/membership/enroll-membership": {
      "post": {
        "tags": [
          "member-sub-controller"
        ],
        "operationId": "enrollMembership",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollMembershipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonEnrollMembershipResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/membership/delete-subscription": {
      "post": {
        "tags": [
          "member-sub-controller"
        ],
        "operationId": "deleteMemberSubscription",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMemberSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonDeleteMemberSubscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/membership/confirm-payment": {
      "post": {
        "tags": [
          "member-sub-controller"
        ],
        "operationId": "enrollMembership_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonPaymentConfirmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/membership/add-subscription": {
      "post": {
        "tags": [
          "member-sub-controller"
        ],
        "operationId": "addMemberSubscription",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonAddMemberSubscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payment/get-payment-url": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getPaymentUrl",
        "parameters": [
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "memberId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonPaymentRes"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/membership/is-user-subscribed": {
      "get": {
        "tags": [
          "member-sub-controller"
        ],
        "operationId": "checkSubscriptionUser",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/membership/check-user-subscription": {
      "get": {
        "tags": [
          "member-sub-controller"
        ],
        "operationId": "checkSubscriptionUser_1",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "subscriptionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonBoolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BenefitDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "UpdateMemberSubscriptionRequest": {
        "type": "object",
        "properties": {
          "subId": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "nameSubscription": {
            "type": "string"
          },
          "fee_member": {
            "type": "number",
            "format": "double"
          },
          "expireDate": {
            "type": "string"
          },
          "maxBook": {
            "type": "integer",
            "format": "int32"
          },
          "selectedBenefits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BenefitDTO"
            }
          }
        }
      },
      "ResponseCommonUpdateMemberSubscriptionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UpdateMemberSubscriptionResponse"
          }
        }
      },
      "UpdateMemberSubscriptionResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "EnrollMembershipRequest": {
        "required": [
          "email",
          "membershipId"
        ],
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "EnrollMembershipResponse": {
        "required": [
          "orderId",
          "urlPayment"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "urlPayment": {
            "type": "string"
          }
        }
      },
      "ResponseCommonEnrollMembershipResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/EnrollMembershipResponse"
          }
        }
      },
      "DeleteMemberSubscriptionRequest": {
        "required": [
          "email",
          "subscriptionId"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "subscriptionId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DeleteMemberSubscriptionResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ResponseCommonDeleteMemberSubscriptionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/DeleteMemberSubscriptionResponse"
          }
        }
      },
      "PaymentConfirmRequest": {
        "required": [
          "orderId",
          "vnp_Amount",
          "vnp_BankCode",
          "vnp_BankTranNo",
          "vnp_CardType",
          "vnp_OrderInfo",
          "vnp_PayDate",
          "vnp_ResponseCode",
          "vnp_SecureHash",
          "vnp_TmnCode",
          "vnp_TransactionNo",
          "vnp_TransactionStatus",
          "vnp_TxnRef"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "vnp_Amount": {
            "type": "string"
          },
          "vnp_BankCode": {
            "type": "string"
          },
          "vnp_BankTranNo": {
            "type": "string"
          },
          "vnp_CardType": {
            "type": "string"
          },
          "vnp_OrderInfo": {
            "type": "string"
          },
          "vnp_PayDate": {
            "type": "string"
          },
          "vnp_ResponseCode": {
            "type": "string"
          },
          "vnp_TmnCode": {
            "type": "string"
          },
          "vnp_TransactionNo": {
            "type": "string"
          },
          "vnp_TransactionStatus": {
            "type": "string"
          },
          "vnp_TxnRef": {
            "type": "string"
          },
          "vnp_SecureHash": {
            "type": "string"
          }
        }
      },
      "PaymentConfirmResponse": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "ResponseCommonPaymentConfirmResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PaymentConfirmResponse"
          }
        }
      },
      "AddMemberSubscriptionRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "nameSubscription": {
            "type": "string"
          },
          "fee_member": {
            "type": "number",
            "format": "double"
          },
          "expireDate": {
            "type": "string"
          },
          "maxBook": {
            "type": "integer",
            "format": "int32"
          },
          "selectedBenefits": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AddMemberSubscriptionResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ResponseCommonAddMemberSubscriptionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AddMemberSubscriptionResponse"
          }
        }
      },
...
      "PaymentRes": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "vnp_TxnRef": {
            "type": "string"
          }
        }
      },
      "ResponseCommonPaymentRes": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PaymentRes"
          }
        }
      },
...
      "ResponseCommonBoolean": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}